double checked that we have prop types on every new component in toolbar
