+++ src/components/toolbar/editor-mode-title.jsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import Helmet from 'react-helmet'
+
 import { ToolbarGroup } from 'material-ui/Toolbar'
 import { connect } from 'react-redux'
