+++ src/components/toolbar/notebook-menu-subsection.jsx
@@ -1,8 +1,13 @@
-import MenuItem from 'material-ui/MenuItem'
 import React from 'react'
+import PropTypes from 'prop-types'
+
+import MenuItem from 'material-ui/MenuItem'
 import ArrowDropRight from 'material-ui/svg-icons/navigation-arrow-drop-right'
 
 export default class NotebookMenuSubsection extends React.Component {
+  static propTypes = {
+    title: PropTypes.string,
+  }
   render() {
     const children = React.Children.map(this.props.children, c => React.cloneElement(c, { className: ' medium-menu', width: 256 }))
     return (
