+++ src/components/toolbar/notebook-task-button.jsx
@@ -1,7 +1,18 @@
 import React from 'react'
+import PropTypes from 'prop-types'
 import IconButton from 'material-ui/IconButton'
 
+import UserTask from '../../user-task'
+import ExternalLinkTask from '../../external-link-task'
+
 export default class NotebookTaskFunction extends React.Component {
+  static propTypes = {
+    title: PropTypes.string,
+    task: PropTypes.oneOfType([
+      PropTypes.instanceOf(UserTask),
+      PropTypes.instanceOf(ExternalLinkTask),
+    ]),
+  }
   render() {
     return (
       <IconButton
